Move grid, topograpy, mask back into setup
