+++ roms_tools/setup/initial_conditions.py
@@ -6,7 +6,7 @@ from typing import Dict, Union, List, Optional
 import matplotlib.pyplot as plt
 from pathlib import Path
 from datetime import datetime
-from roms_tools.grid import Grid
+from roms_tools import Grid
 from roms_tools.vertical_coordinate import compute_depth
 from roms_tools.regrid import LateralRegrid, VerticalRegrid
 from roms_tools.plot import _plot, _section_plot, _profile_plot, _line_plot
